extract PR fetching logic as reusable hook
